+++ src/main/scala/longevity/context/LongevityContextImpl.scala
@@ -2,7 +2,7 @@ package longevity.context
 
 import com.typesafe.config.Config
 import emblem.emblematic.traversors.sync.CustomGeneratorPool
-import longevity.persistence.buildRepoPool
+import longevity.persistence.RepoPoolBuilder.buildRepoPool
 import longevity.subdomain.Subdomain
 
 private[context] final class LongevityContextImpl (