+++ src/main/scala/longevity/persistence/Repo.scala
@@ -1,7 +1,5 @@
 package longevity.persistence
 
-import akka.NotUsed
-import akka.stream.scaladsl.Source
 import longevity.subdomain.KeyVal
 import longevity.subdomain.persistent.Persistent
 import longevity.subdomain.ptype.Query
@@ -47,12 +45,6 @@ trait Repo[P <: Persistent] {
    * @param executionContext the execution context
    */
   def retrieveByQuery(query: Query[P])(implicit executionContext: ExecutionContext): Future[Seq[PState[P]]]
-
-  /** streams persistent objects matching a query
-   * 
-   * @param query the query to execute
-   */
-  def streamByQuery(query: Query[P]): Source[PState[P], NotUsed]
 
   /** updates the persistent object
    *