+++ src/main/scala/longevity/persistence/package.scala
@@ -1,29 +1,34 @@
 package longevity
 
-import com.datastax.driver.core.Session
-import com.mongodb.casbah.MongoDB
-import com.typesafe.config.Config
+import akka.NotUsed
+import akka.stream.scaladsl.Source
 import emblem.TypeKey
-import emblem.TypeKeyMap
-import emblem.typeBound.TypeBoundPair
 import emblem.typeKey
-import longevity.context.Cassandra
-import longevity.context.InMem
-import longevity.context.Mongo
-import longevity.context.PersistenceStrategy
-import longevity.persistence.cassandra.CassandraRepo
-import longevity.persistence.inmem.InMemRepo
-import longevity.persistence.mongo.MongoRepo
-import longevity.subdomain.ptype.DerivedPType
-import longevity.subdomain.ptype.PolyPType
-import longevity.subdomain.Subdomain
+import longevity.subdomain.ptype.Query
 import longevity.subdomain.persistent.Persistent
-import longevity.subdomain.ptype.PType
 import scala.concurrent.Future
 import scala.concurrent.ExecutionContext
 
 /** manages entity persistence operations */
 package object persistence {
+
+  /** provides repository methods that use Akka Streams for repository streaming
+   * API.
+   * 
+   * `StreamingRepo` is provided by an implicit conversion from `Repo`, so that
+   * Akka Streams can remain an optional dependency for longevity users.
+   * otherwise, it would have been included as part of the [[Repo]].
+   */
+  implicit class StreamingRepo[P <: Persistent](repo: Repo[P]) {
+
+    /** streams persistent objects matching a query
+     * 
+     * @param query the query to execute
+     */
+    def streamByQuery(query: Query[P]): Source[PState[P], NotUsed] =
+      repo.asInstanceOf[BaseRepo[P]].streamByQueryImpl(query)
+
+  }
 
   /** packages a [[longevity.subdomain.persistent.Persistent persistent object]]
    * with a `TypeKey` for the object's type. used by [[RepoPool.createMany]].
@@ -86,91 +91,6 @@ package object persistence {
         }
       }
 
-  }
-
-  private[longevity] def buildRepoPool(
-    subdomain: Subdomain,
-    persistenceStrategy: PersistenceStrategy,
-    config: Config)
-  : RepoPool =
-    persistenceStrategy match {
-      case InMem => inMemTestRepoPool(subdomain)
-      case Mongo => mongoRepoPool(subdomain, MongoRepo.mongoDbFromConfig(config))
-      case Cassandra => cassandraRepoPool(subdomain, CassandraRepo.sessionFromConfig(config))
-    }
-
-  private trait StockRepoFactory[R[P <: Persistent] <: BaseRepo[P]] {
-    def build[P <: Persistent](
-      pType: PType[P],
-      polyRepoOpt: Option[R[_ >: P <: Persistent]] = None)
-    : R[P]
-  }
-
-  private def inMemTestRepoPool(subdomain: Subdomain): RepoPool = {
-    object repoFactory extends StockRepoFactory[InMemRepo] {
-      def build[P <: Persistent](
-        pType: PType[P],
-        polyRepoOpt: Option[InMemRepo[_ >: P <: Persistent]])
-      : InMemRepo[P] =
-        InMemRepo[P](pType, subdomain, polyRepoOpt)
-    }
-    buildRepoPool(subdomain, repoFactory)
-  }
-
-  private def mongoRepoPool(subdomain: Subdomain, mongoDB: MongoDB): RepoPool = {
-    object repoFactory extends StockRepoFactory[MongoRepo] {
-      def build[P <: Persistent](
-        pType: PType[P],
-        polyRepoOpt: Option[MongoRepo[_ >: P <: Persistent]])
-      : MongoRepo[P] =
-        MongoRepo[P](pType, subdomain, mongoDB, polyRepoOpt)
-    }
-    buildRepoPool(subdomain, repoFactory)
-  }
-
-  private def cassandraRepoPool(subdomain: Subdomain, session: Session): RepoPool = {
-    object repoFactory extends StockRepoFactory[CassandraRepo] {
-      def build[P <: Persistent](
-        pType: PType[P],
-        polyRepoOpt: Option[CassandraRepo[_ >: P <: Persistent]])
-      : CassandraRepo[P] =
-        CassandraRepo[P](pType, subdomain, session, polyRepoOpt)
-    }
-    buildRepoPool(subdomain, repoFactory)
-  }
-
-  private def buildRepoPool[R[P <: Persistent] <: BaseRepo[P]](
-    subdomain: Subdomain,
-    stockRepoFactory: StockRepoFactory[R])
-  : RepoPool = {
-    var keyToRepoMap = TypeKeyMap[Persistent, R]
-    type Pair[P <: Persistent] = TypeBoundPair[Persistent, TypeKey, PType, P]
-    def createRepoFromPair[P <: Persistent](pair: Pair[P]): Unit = {
-      val pTypeKey = pair._1
-      val pType = pair._2
-
-      val polyKey: Option[TypeKey[_ >: P <: Persistent]] = pType match {
-        case dpt: DerivedPType[_, _] => Some(dpt.polyPType.pTypeKey)
-        case _ => None
-      }
-
-      val repo = stockRepoFactory.build[P](pType, polyKey.map(keyToRepoMap(_)))
-      keyToRepoMap += (pTypeKey -> repo)
-
-    }
-    subdomain.pTypePool.filter(isPolyPType).iterator.foreach { pair => createRepoFromPair(pair) }
-    subdomain.pTypePool.filterNot(isPolyPType).iterator.foreach { pair => createRepoFromPair(pair) }
-    val repoPool = new RepoPool(keyToRepoMap.widen[BaseRepo])
-    finishRepoInitialization(repoPool)
-    repoPool
-  }
-
-  private def isPolyPType(pair: TypeBoundPair[Persistent, TypeKey, PType, _ <: Persistent]): Boolean = {
-    pair._2.isInstanceOf[PolyPType[_]]
-  }
-
-  private def finishRepoInitialization(repoPool: RepoPool): Unit = {
-    repoPool.baseRepoMap.values.foreach { repo => repo._repoPoolOption = Some(repoPool) }
   }
 
 }